make sure indices are ints

+++ circus/files/openephys.py
@@ -72,7 +72,6 @@ class OpenEphysFile(DataFile):
 
 
     def _get_slice_(self, t_start, t_stop):
-
         x_beg = numpy.int64(t_start // self.SAMPLES_PER_RECORD)
         r_beg = numpy.mod(t_start, self.SAMPLES_PER_RECORD)
         x_end = numpy.int64(t_stop // self.SAMPLES_PER_RECORD)
@@ -82,17 +81,17 @@ class OpenEphysFile(DataFile):
 
         if x_beg == x_end:
             g_offset = x_beg * self.SAMPLES_PER_RECORD + self.OFFSET_PER_BLOCK[0]*(x_beg + 1) + self.OFFSET_PER_BLOCK[1]*x_beg
-            data_slice = numpy.arange(g_offset + r_beg, g_offset + r_end)
+            data_slice = numpy.arange(g_offset + r_beg, g_offset + r_end, dtype=numpy.int64)
         else:
-            for count, nb_blocks in enumerate(numpy.arange(x_beg, x_end + 1)):
+            for count, nb_blocks in enumerate(numpy.arange(x_beg, x_end + 1, dtype=numpy.int64)):
                 g_offset = nb_blocks * self.SAMPLES_PER_RECORD + self.OFFSET_PER_BLOCK[0]*(nb_blocks + 1) + self.OFFSET_PER_BLOCK[1]*nb_blocks
                 if count == 0:
-                    data_slice += numpy.arange(g_offset + r_beg, g_offset + self.SAMPLES_PER_RECORD).tolist()
+                    data_slice += numpy.arange(g_offset + r_beg, g_offset + self.SAMPLES_PER_RECORD, dtype=numpy.int64).tolist()
                 elif (count == (x_end - x_beg)):
-                    data_slice += numpy.arange(g_offset, g_offset + r_end).tolist()
+                    data_slice += numpy.arange(g_offset, g_offset + r_end, dtype=numpy.int64).tolist()
                 else:
-                    data_slice += numpy.arange(g_offset, g_offset + self.SAMPLES_PER_RECORD).tolist()
-        return data_slice 
+                    data_slice += numpy.arange(g_offset, g_offset + self.SAMPLES_PER_RECORD, dtype=numpy.int64).tolist()
+        return data_slice
 
 
     def read_chunk(self, idx, chunk_size, padding=(0, 0), nodes=None):
